matrix, pascal triangle add

+++ c/sum_avg.c
@@ -3,12 +3,12 @@ int main()
 {
 	int n = 4;
 	int num[]={4,5,1,2};
-	int sum =0;
+	float sum =0;
 	for (int i = 0; i < n; i++)
 	{
-		sum=sum+num[i];
+		 sum=sum+num[i];
 	}
 	int avg =sum/n;
-	printf("Avg is %d ", avg);
+	printf("Avg is %f ", avg);
 	return 0;
 }
